Complete porting tests to compile on JVM and JS
