+++ tests/jvm/src/test/resources/Animated_PNG_example_bouncing_beach_ball.png
