+++ tests/jvm/src/test/resources/ball.png
