+++ tests/jvm/src/test/resources/lorem-ipsum.txt
