+++ tests/jvm/src/test/scala/org/http4s/StaticFileSpec.scala
@@ -62,9 +62,9 @@ class StaticFileSpec extends Http4sSpec {
       }
 
       val tests = List(
-        "./testing/src/test/resources/logback-test.xml",
+        "./testing/shared/src/test/resources/logback-test.xml",
         "./server/src/test/resources/testresource.txt",
-        ".travis.yml")
+        "./.travis.yml")
 
       forall(tests)(check)
     }
